requirements.txt
pyzbar
flask
flask-cors
pillow
waitress
